Actualizar Printf.h, Printf.c y 6 más archivos...

+++ TM4C129E_ApplicationLib/xUtils/Conversion/Conversion_Number2String/Conversion_Primitives/xSource/Conversion_Inversion.c
@@ -83,7 +83,7 @@ CONV_nSTATUS Conv__enOutInversion(CONV_OUT_t pvfOut, char* pvBufferOut, const ch
         }
 
     }
-    *pu32BufOutLenght = u32Index;
+    *pu32BufOutLenght = (uint32_t) u32Index;
     return (enStatus);
 }
